[Routing] Setup a nested route and <Outlet/> in Home page, to allow dynamic render for displaying projects. Slight UI changes
