+++ src/App.tsx
@@ -3,12 +3,16 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { Home } from "./pages/Home";
 import { About } from "./pages/About";
 import { Projects } from "./pages/Projects";
+import { MainContent } from './components/MainContent';
 
 function App() {
     return (
         <Router>
             <Routes>
-                <Route path='/' element={<Home />}/>
+                <Route path='/' element={<Home />}>
+                    <Route index element={<MainContent />}/>
+                    <Route path='projects/:id'/>
+                </Route>
                 <Route path='/about' element={<About />}/>
                 <Route path='/projects' element={<Projects />}/>
             </Routes>
