+++ src/styles/Home.module.scss
@@ -20,10 +20,17 @@ body, html {
 .header {
     width: 100%;
     text-align: center;
-    padding: 20px;
+    height: 60px;
     font-size: 4rem;
-    background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), 
-    url("./src/assets/header-background.jpg") no-repeat center center/cover;
+    background-color: #1A1A1A;
+}
+
+.siteIcon {
+    font-size: 22px;
+    color: white;
+    text-align: left;
+    padding: 10px;
+    margin: 10px;
 }
 
 .footer {
@@ -174,3 +181,5 @@ button:hover {
      color: white;
      font-size: 20px;
 }
+
+
